adicion de la variable uint32_t en todas las funciones y creacion del archivo text.c

+++ imprimir.c
@@ -1,12 +1,9 @@
 #include "imprimir.h"
-#include <stdio.h>
-void valor_registro( unsigned long int valores[])
+void valor_registro( uint32_t  valores[])
 {
-
-
     int i;
     for(i=0;i<12;i++)
     {
-        printf("Registro %d: %ld\n",i,valores[i]);
+        printf("Registro %d: %d\n",i,valores[i]);
     }
 }
