+++ imprimir.h
@@ -1,1 +1,4 @@
-void valor_registro( unsigned long int valores[]);
+#include <stdio.h>
+#include <stdint.h>
+#include <stdlib.h>
+void valor_registro( uint32_t valores[]);
